// frontend/lib/api/generated/model/recommendedJob.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Career-Hi Backend
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendedJobRecIdx } from './recommendedJobRecIdx'
import type { RecommendedJobDeadline } from './recommendedJobDeadline'
import type { RecommendedJobStartDate } from './recommendedJobStartDate'
import type { RecommendedJobCrawlingTime } from './recommendedJobCrawlingTime'

/**
 * 추천된 채용공고
 */
export interface RecommendedJob {
  rec_idx?: RecommendedJobRecIdx
  title: string
  url: string
  deadline?: RecommendedJobDeadline
  start_date?: RecommendedJobStartDate
  crawling_time?: RecommendedJobCrawlingTime
  recommendation_reason: string
}